refactor: use toolbox for version and jwt validation
