+++ pkg/api/v1/router.go
@@ -8,9 +8,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 	"github.com/volatiletech/sqlboiler/v4/boil"
+	"go.hollow.sh/toolbox/ginjwt"
 
 	"go.hollow.sh/dcim/internal/models"
-	"go.hollow.sh/dcim/pkg/ginjwt"
 )
 
 // Router provides a router for the v1 API
