--- pkg/ginjwt/doc.go
@@ -1,2 +1,0 @@
-// Package ginjwt provides a JWT authentication and authorization middleware for use with a gin server
-package ginjwt
