--- pkg/version/doc.go
@@ -1,3 +1,0 @@
-// Package version provides version strings and version information for the
-// application.
-package version // import "go.hollow.sh/dcim/pkg/version"
